Update version to 1.2.3027, enhance documentation, and improve UI scaling handling

+++ src/studio/system.h
@@ -147,7 +147,7 @@ typedef struct
 
     const tic_cartridge* cart;
 
-    s32 uiScale;
+    float uiScale;
 
     int fft;
     int fftcaptureplaybackdevices;
